requirements.txt
streamlit>=1.28.0
langchain>=0.3.0
langchain-openai>=0.3.0
langchain-community>=0.3.0
chromadb>=0.5.0
python-dotenv>=1.0.0
pandas>=2.0.0
requests>=2.31.0
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

